Udemy/MLOps/Sources/mlops_project_final/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ml_classifier"
version = "0.4"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow==2.10.5",
    "dagshub==0.5.9",
    "dvc==3.59.1",
    "pathspec<1.0",
    "flask==2.2.5",
    "gunicorn==23.0.0",
    "mlflow==2.22.1",
    "numpy==2.1.3",
    "pandas==2.2.3",
    "python-dotenv==1.1.0",
    "PyYAML==6.0.2",
    "scikit-learn==1.6.1",
    "tensorflow==2.19.0"
]

[tool.setuptools.packages.find]
where = ["."]
